Prev button....not really working.
It is more involved than planned. Will need to undo the previous snapshot's modification of the galaxy to restore the original, for which I will need to replay the galaxy state.

Alternatively, I create a stack of the galaxy to allow for easy forwards/backwards, at the expense of more memory.