New history file format for recording the history of a game.
New format has the base galaxy, the current state of the galaxy, and snapshots with full undo history to enable easy forwards/backwards movement on the timeline.

Updates to the galaxy now come in through `/incr_update`, which sends individual updates to the galaxy. Note: sometimes we get superfluous updates from the game itself that don't seem to indicate any change. Still investigating this.

The old snapshot history is now obsolete. It will slowly be replaced and removed.